removing services from app
