+++ src/app/app.component.html
@@ -1,5 +1,5 @@
 <div class="container">
-  <app-time-display time="{{ time }}"></app-time-display>
+  <app-time-display time="{{ formatTime }}"></app-time-display>
   <div class="btn-container">
     <app-control-button
       icon="play_arrow"
