+++ src/app/app.component.ts
@@ -3,9 +3,9 @@
  */
 
 import { Component } from '@angular/core';
-import { Subscription } from 'rxjs';
-
-import { TimerService } from './services/timer.service';
+import { Subject, Subscription, fromEvent, timer } from 'rxjs';
+import { takeUntil, map, filter, buffer, debounceTime } from 'rxjs/operators';
+import { convertTime } from './utils/utils';
 
 @Component({
   selector: 'app-root',
@@ -19,29 +19,63 @@ export class AppComponent {
   /** the variable is assigned a value after subscribing to the timer after start */
   subscription!: Subscription;
 
-  time?: string;
+  /** time in HH:MM:SS format which is string type*/
+  formatTime?: string;
 
-  constructor(private timerService: TimerService) {}
+  /** amount of time in seconds */
+  time: number = 0;
+  /** emit event with stop trggier */
+  private stop$ = new Subject<void>();
 
-  /**Starts a timer by subscribing */
+  constructor() {}
+
+  /**
+   * Start the timer from the time it was stopped
+   * The method monitors and assigns a new value to the "time" variable every second
+   * and returns time in HH:MM:SS format
+   *
+   */
   startTimer(): void {
-    this.subscription = this.timerService
-      .start()
-      .subscribe((value) => (this.time = value));
+    this.subscription = timer(0, 1000)
+      .pipe(
+        map(() => {
+          this.time++;
+          return convertTime(this.time);
+        }),
+        takeUntil(this.stop$)
+      )
+      .subscribe((value) => (this.formatTime = value));
   }
 
   /**Stop a timer by double click with an interval of less than 300 ms*/
   waitTimer(): void {
-    this.timerService.wait();
+    const waitBtn = document.getElementById('waitBtn');
+    if (waitBtn) {
+      const buttonClick$ = fromEvent(waitBtn, 'click');
+
+      // tracks double clicks with an interval of less than 300 ms, then stop the timer
+      buttonClick$
+        .pipe(
+          buffer(buttonClick$.pipe(debounceTime(300))),
+          filter((clicks) => clicks.length >= 2)
+        )
+        .subscribe(() => this.stop$.next());
+    }
   }
 
-  /**Stop a timer */
+  /**  Stop a timer
+   *   tracks double clicks on the "waitBtn" button to stop the timer
+   */
   stopTimer(): void {
-    this.timerService.stop();
+    this.stop$.next();
   }
 
   /**Reset timer's value to 00:00:00 */
   resetTimer(): void {
-    this.time = this.timerService.reset();
+    this.stopTimer();
+
+    // Reset time property in order to start over
+    this.time = 0;
+    this.formatTime = convertTime(this.time);
   }
 }
